+++ geode-core/src/main/java/org/apache/geode/cache/client/internal/ConnectionImpl.java
@@ -36,8 +36,8 @@ import org.apache.geode.cache.wan.GatewaySender;
 import org.apache.geode.distributed.internal.DistributionConfig;
 import org.apache.geode.distributed.internal.InternalDistributedSystem;
 import org.apache.geode.distributed.internal.ServerLocation;
+import org.apache.geode.internal.cache.tier.ClientSideHandshake;
 import org.apache.geode.internal.cache.tier.CommunicationMode;
-import org.apache.geode.internal.cache.tier.sockets.HandShake;
 import org.apache.geode.internal.cache.tier.sockets.ServerConnection;
 import org.apache.geode.internal.cache.tier.sockets.ServerQueueStatus;
 import org.apache.geode.internal.i18n.LocalizedStrings;
@@ -87,7 +87,7 @@ public class ConnectionImpl implements Connection {
 
   private long connectionID = Connection.DEFAULT_CONNECTION_ID;
 
-  private HandShake handShake;
+  private ClientSideHandshake handshake;
 
   public ConnectionImpl(InternalDistributedSystem ds, CancelCriterion cancelCriterion) {
     this.ds = ds;
@@ -100,10 +100,10 @@ public class ConnectionImpl implements Connection {
   }
 
   public ServerQueueStatus connect(EndpointManager endpointManager, ServerLocation location,
-      HandShake handShake, int socketBufferSize, int handShakeTimeout, int readTimeout,
+      ClientSideHandshake handshake, int socketBufferSize, int handshakeTimeout, int readTimeout,
       CommunicationMode communicationMode, GatewaySender sender, SocketCreator sc)
       throws IOException {
-    theSocket = sc.connectForClient(location.getHostName(), location.getPort(), handShakeTimeout,
+    theSocket = sc.connectForClient(location.getHostName(), location.getPort(), handshakeTimeout,
         socketBufferSize);
     theSocket.setTcpNoDelay(true);
     // System.out.println("ConnectionImpl setting buffer sizes: " +
@@ -114,10 +114,10 @@ public class ConnectionImpl implements Connection {
     verifySocketBufferSize(socketBufferSize, theSocket.getReceiveBufferSize(), "receive");
     verifySocketBufferSize(socketBufferSize, theSocket.getSendBufferSize(), "send");
 
-    theSocket.setSoTimeout(handShakeTimeout);
+    theSocket.setSoTimeout(handshakeTimeout);
     out = theSocket.getOutputStream();
     in = theSocket.getInputStream();
-    this.status = handShake.handshakeWithServer(this, location, communicationMode);
+    this.status = handshake.handshakeWithServer(this, location, communicationMode);
     commBuffer = ServerConnection.allocateCommBuffer(socketBufferSize, theSocket);
     if (sender != null) {
       commBufferForAsyncRead = ServerConnection.allocateCommBuffer(socketBufferSize, theSocket);
@@ -310,12 +310,16 @@ public class ConnectionImpl implements Connection {
     return this.connectionID;
   }
 
-  protected HandShake getHandShake() {
-    return handShake;
+  protected byte[] encryptBytes(byte[] messageBytes) throws Exception {
+    return handshake.getEncryptor().encryptBytes(messageBytes);
   }
 
-  protected void setHandShake(HandShake handShake) {
-    this.handShake = handShake;
+  protected byte[] decryptBytes(byte[] messageBytes) throws Exception {
+    return handshake.getEncryptor().decryptBytes(messageBytes);
+  }
+
+  protected void setHandshake(ClientSideHandshake handshake) {
+    this.handshake = handshake;
   }
 
   /**
