+++ geode-core/src/main/java/org/apache/geode/internal/cache/properties.html
@@ -271,13 +271,13 @@ This is the TCP accept backlog for the acceptor thread's listening socket.
 </dd>
 
 <!-- -------------------------------------------------------  -->
-<dt><strong>BridgeServer.handShakeTimeout</strong></dt>
+<dt><strong>BridgeServer.handshakeTimeout</strong></dt>
 <dd>
 <em>Public:</em> false
 <p>
 <em>Integer</em> (default is 59000)
 <p>
-See <code>org.apache.geode.internal.cache.tier.sockets.AcceptorImpl#handShakeTimeout</code>.
+See <code>org.apache.geode.internal.cache.tier.sockets.AcceptorImpl#handshakeTimeout</code>.
 <p>
 <pre>
   Test value for handshake timeout
