+++ geode-core/src/main/java/org/apache/geode/internal/cache/tier/ClientSideHandshake.java
@@ -1,0 +1,40 @@
+/*
+ * Licensed to the Apache Software Foundation (ASF) under one or more contributor license
+ * agreements. See the NOTICE file distributed with this work for additional information regarding
+ * copyright ownership. The ASF licenses this file to You under the Apache License, Version 2.0 (the
+ * "License"); you may not use this file except in compliance with the License. You may obtain a
+ * copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
+package org.apache.geode.internal.cache.tier;
+
+import java.io.IOException;
+import java.net.Socket;
+
+import org.apache.geode.cache.client.ServerRefusedConnectionException;
+import org.apache.geode.cache.client.internal.Connection;
+import org.apache.geode.distributed.internal.ServerLocation;
+import org.apache.geode.internal.cache.tier.sockets.ClientProxyMembershipID;
+import org.apache.geode.internal.cache.tier.sockets.ServerQueueStatus;
+import org.apache.geode.security.AuthenticationFailedException;
+import org.apache.geode.security.AuthenticationRequiredException;
+
+public interface ClientSideHandshake {
+  Encryptor getEncryptor();
+
+  ClientProxyMembershipID getMembershipId();
+
+  ServerQueueStatus handshakeWithSubscriptionFeed(Socket socket, boolean isPrimary)
+      throws IOException, AuthenticationRequiredException, AuthenticationFailedException,
+      ServerRefusedConnectionException, ClassNotFoundException;
+
+  ServerQueueStatus handshakeWithServer(Connection conn, ServerLocation location,
+      CommunicationMode communicationMode) throws IOException, AuthenticationRequiredException,
+      AuthenticationFailedException, ServerRefusedConnectionException;
+}
