+++ geode-core/src/main/java/org/apache/geode/internal/cache/tier/sockets/MessageIdExtractor.java
@@ -15,11 +15,13 @@
  */
 package org.apache.geode.internal.cache.tier.sockets;
 
+import org.apache.geode.internal.cache.tier.Encryptor;
+import org.apache.geode.internal.cache.tier.ServerSideHandshake;
 import org.apache.geode.internal.i18n.LocalizedStrings;
 import org.apache.geode.security.AuthenticationRequiredException;
 
 public class MessageIdExtractor {
-  public long getUniqueIdFromMessage(Message requestMessage, HandShake handshake, long connectionId)
+  public long getUniqueIdFromMessage(Message requestMessage, Encryptor handshake, long connectionId)
       throws AuthenticationRequiredException {
     AuthIds aIds = getAuthIdsFromMessage(requestMessage, handshake);
     if (connectionId != aIds.getConnectionId()) {
@@ -29,7 +31,7 @@ public class MessageIdExtractor {
     return aIds.getUniqueId();
   }
 
-  private AuthIds getAuthIdsFromMessage(Message requestMessage, HandShake handshake)
+  private AuthIds getAuthIdsFromMessage(Message requestMessage, Encryptor handshake)
       throws AuthenticationRequiredException {
     try {
       byte[] secureBytes = requestMessage.getSecureBytes();
