+++ geode-core/src/main/java/org/apache/geode/internal/cache/tier/sockets/command/ExecuteRegionFunction.java
@@ -16,7 +16,6 @@ package org.apache.geode.internal.cache.tier.sockets.command;
 
 import java.io.IOException;
 import java.util.HashSet;
-import java.util.Optional;
 import java.util.Set;
 
 import org.apache.geode.cache.Region;
@@ -35,9 +34,9 @@ import org.apache.geode.internal.cache.execute.ServerToClientFunctionResultSende
 import org.apache.geode.internal.cache.tier.CachedRegionHelper;
 import org.apache.geode.internal.cache.tier.Command;
 import org.apache.geode.internal.cache.tier.MessageType;
+import org.apache.geode.internal.cache.tier.ServerSideHandshake;
 import org.apache.geode.internal.cache.tier.sockets.BaseCommand;
 import org.apache.geode.internal.cache.tier.sockets.ChunkedMessage;
-import org.apache.geode.internal.cache.tier.sockets.HandShake;
 import org.apache.geode.internal.cache.tier.sockets.Message;
 import org.apache.geode.internal.cache.tier.sockets.Part;
 import org.apache.geode.internal.cache.tier.sockets.ServerConnection;
@@ -137,9 +136,9 @@ public class ExecuteRegionFunction extends BaseCommand {
       return;
     }
 
-    HandShake handShake = (HandShake) serverConnection.getHandshake();
-    int earlierClientReadTimeout = handShake.getClientReadTimeout();
-    handShake.setClientReadTimeout(0);
+    ServerSideHandshake handshake = serverConnection.getHandshake();
+    int earlierClientReadTimeout = handshake.getClientReadTimeout();
+    handshake.setClientReadTimeout(0);
     ServerToClientFunctionResultSender resultSender = null;
     Function<?> functionObject = null;
     try {
@@ -240,7 +239,7 @@ public class ExecuteRegionFunction extends BaseCommand {
       String message = e.getMessage();
       sendException(hasResult, clientMessage, message, serverConnection, e);
     } finally {
-      handShake.setClientReadTimeout(earlierClientReadTimeout);
+      handshake.setClientReadTimeout(earlierClientReadTimeout);
     }
   }
