+++ geode-core/src/test/java/org/apache/geode/internal/cache/tier/sockets/BackwardCompatibilityHigherVersionClientDUnitTest.java
@@ -30,6 +30,7 @@ import org.apache.geode.cache.Region;
 import org.apache.geode.cache.RegionAttributes;
 import org.apache.geode.cache.Scope;
 import org.apache.geode.cache.client.PoolManager;
+import org.apache.geode.cache.client.internal.ClientSideHandshakeImpl;
 import org.apache.geode.cache.client.internal.ConnectionFactoryImpl;
 import org.apache.geode.cache.client.internal.PoolImpl;
 import org.apache.geode.cache.server.CacheServer;
@@ -169,7 +170,7 @@ public class BackwardCompatibilityHigherVersionClientDUnitTest extends JUnit4Dis
    * handshake.
    */
   public static void setHandshakeVersionForTesting() throws Exception {
-    HandShake.setVersionForTesting((short) (currentClientVersion + 1));
+    ClientSideHandshakeImpl.setVersionForTesting((short) (currentClientVersion + 1));
   }
 
   public static void addExceptions() throws Exception {
@@ -211,7 +212,7 @@ public class BackwardCompatibilityHigherVersionClientDUnitTest extends JUnit4Dis
    * handshake.
    */
   public static void unsetHandshakeVersionForTesting() throws Exception {
-    HandShake.setVersionForTesting(currentClientVersion);
+    ClientSideHandshakeImpl.setVersionForTesting(currentClientVersion);
   }
 
   /*
